backend/requirements.txt
beautifulsoup4==4.12.3
certstream==1.12
Django==6.0.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
ioc-finder==7.3.0
ipwhois==1.3.0
Levenshtein==0.27.3
opencv-python==4.13.0.92
openpyxl==3.1.5
pandas==2.3.2
psycopg2-binary==2.9.10
pulsedive==0.1.0
pytesseract==0.3.13
python-dotenv==1.1.1
python-whois==0.9.6
requests==2.32.5
selenium==4.27.1
tldextract==5.3.1
Unidecode==1.4.0
waitress==3.0.2
